Regex substitution and shell utilities
