+++ .gitignore
@@ -1,2 +1,4 @@
 __pycache__/
 .mypy_cache/
+debian/helper-scripts-tests/
+debian/helper-scripts/
