+++ usr/bin/stcat
@@ -1,0 +1,31 @@
+#!/usr/bin/env python3
+
+## SPDX-FileCopyrightText: 2025 Benjamin Grande M. S. <ben.grande.b@gmail.com>
+## SPDX-FileCopyrightText: 2025 ENCRYPTED SUPPORT LLC <adrelanos@whonix.org>
+##
+## SPDX-License-Identifier: AGPL-3.0-or-later
+
+"""Safely print stdin or file to stdout."""
+
+from fileinput import input as file_input
+from sys import argv, stdin, stdout
+from stdisplay.stdisplay import stdisplay
+
+
+def main() -> None:
+    """Safely print stdin or file to stdout."""
+    if len(argv) > 1:
+        ## File input can be '-' (stdin).
+        stdin.reconfigure(errors="ignore")  # type: ignore
+        for untrusted_text in file_input(encoding="ascii"):
+            stdout.write(stdisplay(untrusted_text))
+        stdout.flush()
+    else:
+        stdin.reconfigure(errors="ignore")  # type: ignore
+        for untrusted_text in stdin:
+            stdout.write(stdisplay(untrusted_text))
+        stdout.flush()
+
+
+if __name__ == "__main__":
+    main()
