+++ usr/bin/stsponge
@@ -1,0 +1,29 @@
+#!/usr/bin/env python3
+
+## SPDX-FileCopyrightText: 2025 Benjamin Grande M. S. <ben.grande.b@gmail.com>
+## SPDX-FileCopyrightText: 2025 ENCRYPTED SUPPORT LLC <adrelanos@whonix.org>
+##
+## SPDX-License-Identifier: AGPL-3.0-or-later
+
+"""Safely print stdin to stdout or file."""
+
+from sys import argv, stdin, stdout
+from stdisplay.stdisplay import stdisplay
+
+
+def main() -> None:
+    """Safely print stdin to stdout or file."""
+    untrusted_text_list = []
+    stdin.reconfigure(errors="ignore")  # type: ignore
+    for untrusted_text in stdin:
+        untrusted_text_list.append(untrusted_text)
+    if len(argv) == 1:
+        stdout.write(stdisplay("".join(untrusted_text_list)))
+        stdout.flush()
+    else:
+        with open(argv[1], mode="w", encoding="ascii") as file:
+            file.write(stdisplay("".join(untrusted_text_list)))
+
+
+if __name__ == "__main__":
+    main()
