+++ usr/bin/sttee
@@ -1,0 +1,27 @@
+#!/usr/bin/env python3
+
+## SPDX-FileCopyrightText: 2025 Benjamin Grande M. S. <ben.grande.b@gmail.com>
+## SPDX-FileCopyrightText: 2025 ENCRYPTED SUPPORT LLC <adrelanos@whonix.org>
+##
+## SPDX-License-Identifier: AGPL-3.0-or-later
+
+"""Safely print stdin to stdout and file."""
+from sys import argv, stdin, stdout
+from stdisplay.stdisplay import stdisplay
+
+
+def main() -> None:
+    """Safely print stdin to stdout and file."""
+    untrusted_text = ""
+    stdin.reconfigure(errors="ignore")  # type: ignore
+    for untrusted_text in stdin:
+        stdout.write(stdisplay(untrusted_text))
+    stdout.flush()
+    if len(argv) > 1:
+        for file_arg in argv[1:]:
+            with open(file_arg, mode="w", encoding="ascii") as file:
+                file.write(stdisplay(untrusted_text))
+
+
+if __name__ == "__main__":
+    main()
